+++ shop/models_bases/managers.py
@@ -68,7 +68,7 @@ class OrderManager(models.Manager):
             return None
 
     @transaction.commit_on_success
-    def create_from_cart(self, cart):
+    def create_from_cart(self, cart, state=None):
         """
         This creates a new Order object (and all the rest) from a passed Cart
         object.
@@ -79,6 +79,10 @@ class OrderManager(models.Manager):
         This will only actually commit the transaction once the function exits
         to minimize useless database access.
 
+        The `state` parameter is further passed to process_cart_item,
+        process_cart, and post_process_cart, so it can be used as a way to
+        store per-request arbitrary information.
+
         Emits the ``processing`` signal.
         """
         # must be imported here!
@@ -88,14 +92,9 @@ class OrderManager(models.Manager):
             OrderItem,
         )
         from shop.models.cartmodel import CartItem
-        # Let's create the Order itself:
-        order = self.model()
-        order.user = cart.user
-        order.status = self.model.PROCESSING  # Processing
 
-        order.order_subtotal = cart.subtotal_price
-        order.order_total = cart.total_price
-
+        # Create an empty order object
+        order = self.create_order_object(cart, state)
         order.save()
 
         # Let's serialize all the extra price arguments in DB
@@ -109,7 +108,7 @@ class OrderManager(models.Manager):
         # There, now move on to the order items.
         cart_items = CartItem.objects.filter(cart=cart)
         for item in cart_items:
-            item.update(cart)
+            item.update(state)
             order_item = OrderItem()
             order_item.order = order
             order_item.product_reference = item.product.get_product_reference()
@@ -130,4 +129,15 @@ class OrderManager(models.Manager):
                 eoi.save()
 
         processing.send(self.model, order=order, cart=cart)
+        return order
+
+    def create_order_object(self, cart, state):
+        """
+        Create an empty order object and fill it with the given cart data.
+        """
+        order = self.model()
+        order.user = cart.user
+        order.status = self.model.PROCESSING  # Processing
+        order.order_subtotal = cart.subtotal_price
+        order.order_total = cart.total_price
         return order