+++ CHANGELOG.rst
@@ -1,8 +1,12 @@
+Version NEXT
+============
+
+* 
+
 Version 0.1.1
 =============
 
 * Changed CurrencyField default decimal precision back to 2
-
 
 Version 0.1.0
 =============